+++ EssentialFeed/EssentialFeed/FeedCache/Infrastructure/CoreData/CoreDataFeedStore.swift
@@ -64,59 +64,3 @@ public class CoreDataFeedStore: FeedStore {
         context.perform { action(context) }
     }
 }
-
-@objc(ManagedCache)
-private class ManagedCache: NSManagedObject {
-    @NSManaged var timestamp: Date
-    @NSManaged var feed: NSOrderedSet
-    
-    var localFeed: [LocalFeedImage] {
-        return feed
-            .compactMap { $0 as? ManagedFeedImage }
-            .map(\.local)
-    }
-    
-    static func find(in context: NSManagedObjectContext) throws -> ManagedCache? {
-        let request = NSFetchRequest<ManagedCache>(entityName: ManagedCache.entity().name!)
-        request.returnsObjectsAsFaults = false
-        
-        return try context.fetch(request).first
-    }
-    
-    static func newUniqueInstance(in context: NSManagedObjectContext) throws -> ManagedCache {
-        try deleteFound(in: context)
-        return ManagedCache(context: context)
-    }
-    
-    static func deleteFound(in context: NSManagedObjectContext) throws {
-        try find(in: context).map(context.delete)
-    }
-}
-
-@objc(ManagedFeedImage)
-private class ManagedFeedImage: NSManagedObject {
-    @NSManaged var id: UUID
-    @NSManaged var imageDescription: String?
-    @NSManaged var location: String?
-    @NSManaged var url: URL
-    @NSManaged var cache: ManagedCache
-    
-    var local: LocalFeedImage {
-        return LocalFeedImage(id: id,
-                              description: imageDescription,
-                              location: location,
-                              url: url)
-    }
-    
-    static func images(from localFeed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
-        return NSOrderedSet(array: localFeed.map { local in
-            let managedFeedImage = ManagedFeedImage(context: context)
-            managedFeedImage.id = local.id
-            managedFeedImage.imageDescription = local.description
-            managedFeedImage.location = local.location
-            managedFeedImage.url = local.url
-            
-            return managedFeedImage
-        })
-    }
-}
