+++ EssentialFeed/EssentialFeed/FeedCache/Infrastructure/CoreData/ManagedFeedImage.swift
@@ -1,0 +1,36 @@
+//
+//  ManagedFeedImage.swift
+//  EssentialFeed
+//
+//  Created by Radosław Serek on 25/11/2022.
+//
+
+import CoreData
+
+@objc(ManagedFeedImage)
+class ManagedFeedImage: NSManagedObject {
+    @NSManaged var id: UUID
+    @NSManaged var imageDescription: String?
+    @NSManaged var location: String?
+    @NSManaged var url: URL
+    @NSManaged var cache: ManagedCache
+    
+    var local: LocalFeedImage {
+        return LocalFeedImage(id: id,
+                              description: imageDescription,
+                              location: location,
+                              url: url)
+    }
+    
+    static func images(from localFeed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
+        return NSOrderedSet(array: localFeed.map { local in
+            let managedFeedImage = ManagedFeedImage(context: context)
+            managedFeedImage.id = local.id
+            managedFeedImage.imageDescription = local.description
+            managedFeedImage.location = local.location
+            managedFeedImage.url = local.url
+            
+            return managedFeedImage
+        })
+    }
+}
